pdb.py renamed
Having a pdb.py creates conflicts with python's pbd, Python DeBugger
